feat(alert): refactor DismissButton component and update its stories
